app.jsのリファクタリングを行い、Swagger設定をconfigファイルに移動しました。また、イベントルートを新たに作成し、APIのルーティングを整理しました。これにより、コードの可読性と保守性が向上しました。
