+++ api/routes/events.js
@@ -1,0 +1,262 @@
+const express = require('express');
+const router = express.Router();
+const onchainService = require('../services/onchainService');
+const offchainService = require('../services/offchainService');
+const METRICS = require('../metrics');
+const { createHashBySortedKeys } = require('../util');
+
+/**
+ * @swagger
+ * /api/part-event:
+ *   post:
+ *     summary: 部品イベントを登録
+ *     tags: [Events]
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             $ref: '#/components/schemas/Event'
+ *     responses:
+ *       200:
+ *         description: イベント登録成功
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 success:
+ *                   type: boolean
+ *                 txResult:
+ *                   type: string
+ *       400:
+ *         $ref: '#/components/responses/Error'
+ *       500:
+ *         $ref: '#/components/responses/Error'
+ */
+router.post('/part-event', async (req, res) => {
+  const endTimer = METRICS.transactionDuration.startTimer();
+  METRICS.requestCounter.inc({ method: req.method, route: req.originalUrl });
+
+  const {
+    event_id: eventId,
+    part_id: partId,
+    status,
+    timestamp,
+    location,
+    operator_id: operatorId
+  } = req.body;
+
+  try {
+    const partData = offchainService.getPartById(partId);
+    const supplierData = partData ? offchainService.getSupplierById(partData.supplier_id) : null;
+    const operatorData = offchainService.getOperatorById(operatorId);
+
+    if (!partData || !supplierData || !operatorData) {
+      METRICS.errorCounter.inc({ operation: 'recordPartEvent', error_type: 'missing_master_data' });
+      return res.status(400).json({ error: 'マスターデータが不足しています' });
+    }
+
+    const partHash = createHashBySortedKeys(partData);
+    const supplierHash = createHashBySortedKeys(supplierData);
+    const operatorHash = createHashBySortedKeys(operatorData);
+
+    const txResult = await onchainService.recordPartEvent([
+      eventId, partId, status, timestamp, location, operatorId,
+      partHash, supplierHash, operatorHash
+    ]);
+
+    const newEventData = {
+      event_id: eventId,
+      part_id: partId,
+      status,
+      timestamp,
+      location,
+      operator_id: operatorId,
+      part_hash: partHash,
+      supplier_hash: supplierHash,
+      operator_hash: operatorHash
+    };
+    offchainService.saveEvent(newEventData);
+
+    METRICS.transactionCounter.inc({ operation: 'recordPartEvent', status: 'success' });
+    endTimer({ operation: 'recordPartEvent' });
+
+    res.json({ success: true, txResult: txResult.toString() });
+  } catch (error) {
+    METRICS.transactionCounter.inc({ operation: 'recordPartEvent', status: 'failure' });
+    METRICS.errorCounter.inc({ operation: 'recordPartEvent', error_type: error.name });
+    endTimer({ operation: 'recordPartEvent' });
+    console.error(error);
+    res.status(500).json({ error: error.message });
+  }
+});
+
+/**
+ * @swagger
+ * /api/query/{event_id}:
+ *   get:
+ *     summary: チェーン上からイベントを取得
+ *     tags: [Events]
+ *     parameters:
+ *       - in: path
+ *         name: event_id
+ *         required: true
+ *         schema:
+ *           type: string
+ *         description: 取得するイベントのID
+ *     responses:
+ *       200:
+ *         description: イベントデータ
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 event_id:
+ *                   type: string
+ *                 result:
+ *                   type: object
+ *       500:
+ *         $ref: '#/components/responses/Error'
+ */
+router.get('/query/:event_id', async (req, res) => {
+  const endTimer = METRICS.transactionDuration.startTimer();
+  METRICS.requestCounter.inc({ method: req.method, route: req.originalUrl });
+
+  const { event_id: eventId } = req.params;
+  try {
+    const result = await onchainService.queryPartEvent(eventId);
+    const chainData = JSON.parse(result.toString());
+    METRICS.transactionCounter.inc({ operation: 'queryPartEvent', status: 'success' });
+    endTimer({ operation: 'queryPartEvent' });
+
+    res.json({ event_id: eventId, result: chainData });
+  } catch (error) {
+    METRICS.transactionCounter.inc({ operation: 'queryPartEvent', status: 'failure' });
+    METRICS.errorCounter.inc({ operation: 'queryPartEvent', error_type: error.name });
+    endTimer({ operation: 'queryPartEvent' });
+    console.error(error);
+    res.status(500).json({ error: error.message });
+  }
+});
+
+/**
+ * @swagger
+ * /api/verify/{event_id}:
+ *   get:
+ *     summary: イベントのハッシュを検証
+ *     tags: [Events]
+ *     parameters:
+ *       - in: path
+ *         name: event_id
+ *         required: true
+ *         schema:
+ *           type: string
+ *         description: 検証するイベントのID
+ *     responses:
+ *       200:
+ *         description: 検証結果
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 event_id:
+ *                   type: string
+ *                 isValid:
+ *                   type: boolean
+ *                 onChainHash:
+ *                   type: object
+ *                 offChainHash:
+ *                   type: object
+ *       404:
+ *         $ref: '#/components/responses/Error'
+ *       500:
+ *         $ref: '#/components/responses/Error'
+ */
+router.get('/verify/:event_id', async (req, res) => {
+  const endTimer = METRICS.transactionDuration.startTimer();
+  METRICS.requestCounter.inc({ method: req.method, route: req.originalUrl });
+
+  const { event_id: eventId } = req.params;
+  try {
+    const onChainResult = await onchainService.queryPartEvent(eventId);
+    const onChainData = JSON.parse(onChainResult.toString());
+
+    const offChainData = offchainService.getEventById(eventId);
+    if (!offChainData) {
+      METRICS.errorCounter.inc({ operation: 'verifyPartEvent', error_type: 'event_not_found' });
+      return res.status(404).json({ error: 'オフチェーンイベントが見つかりません' });
+    }
+
+    const isMatch = (
+      onChainData.part_hash === offChainData.part_hash &&
+      onChainData.supplier_hash === offChainData.supplier_hash &&
+      onChainData.operator_hash === offChainData.operator_hash
+    );
+
+    METRICS.transactionCounter.inc({ operation: 'verifyPartEvent', status: 'success' });
+    endTimer({ operation: 'verifyPartEvent' });
+
+    res.json({
+      event_id: eventId,
+      isValid: isMatch,
+      onChainHash: {
+        part_hash: onChainData.part_hash,
+        supplier_hash: onChainData.supplier_hash,
+        operator_hash: onChainData.operator_hash
+      },
+      offChainHash: {
+        part_hash: offChainData.part_hash,
+        supplier_hash: offChainData.supplier_hash,
+        operator_hash: offChainData.operator_hash
+      }
+    });
+  } catch (error) {
+    METRICS.transactionCounter.inc({ operation: 'verifyPartEvent', status: 'failure' });
+    METRICS.errorCounter.inc({ operation: 'verifyPartEvent', error_type: error.name });
+    endTimer({ operation: 'verifyPartEvent' });
+    console.error(error);
+    res.status(500).json({ error: error.message });
+  }
+});
+
+/**
+ * @swagger
+ * /api/events:
+ *   get:
+ *     summary: 全イベントを取得
+ *     tags: [Events]
+ *     responses:
+ *       200:
+ *         description: イベント一覧
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: array
+ *               items:
+ *                 $ref: '#/components/schemas/Event'
+ *       500:
+ *         $ref: '#/components/responses/Error'
+ */
+router.get('/events', (req, res) => {
+  const endTimer = METRICS.transactionDuration.startTimer();
+  METRICS.requestCounter.inc({ method: req.method, route: req.originalUrl });
+
+  try {
+    const events = offchainService.getAllEvents();
+    METRICS.transactionCounter.inc({ operation: 'getAllEvents', status: 'success' });
+    endTimer({ operation: 'getAllEvents' });
+
+    res.json(events);
+  } catch (error) {
+    METRICS.transactionCounter.inc({ operation: 'getAllEvents', status: 'failure' });
+    METRICS.errorCounter.inc({ operation: 'getAllEvents', error_type: error.name });
+    endTimer({ operation: 'getAllEvents' });
+    console.error(error);
+    res.status(500).json({ error: error.message });
+  }
+});
+
+module.exports = router;
